# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Drinkly App",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Your daily water consumption calendar"
  },
  "servers": [
    {
      "url": "https://water-app-backend.onrender.com/"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/water": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get all water",
        "operationId": "getAllWater",
        "description": "Retrieve all water records for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved water records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/water"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Water"
        ],
        "summary": "Add water",
        "operationId": "addWater",
        "description": "Add a new water record for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "5f4a5f6b7e5c7d8f9a6b7c8d"
                  },
                  "amount": {
                    "type": "number",
                    "example": 2000
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-12-01T12:00:00.000Z"
                  },
                  "time": {
                    "type": "string",
                    "example": "12:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Water record added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/water/{_id}": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get water by Id",
        "operationId": "getWaterById",
        "description": "Retrieve all water records for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "_id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved water records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully retrieved water records"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "$ref": "#/components/schemas/water"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "Water"
        ],
        "summary": "Update water",
        "operationId": "updateWater",
        "description": "Update a water record for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "_id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "5f4a5f6b7e5c7d8f9a6b7c8d"
                  },
                  "waterVolume": {
                    "type": "number",
                    "example": 2000
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-12-01T12:00:00.000Z"
                  },
                  "time": {
                    "type": "string",
                    "example": "12:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Water record updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully retrieved water records"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "$ref": "#/components/schemas/water"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Water"
        ],
        "summary": "Delete water",
        "operationId": "deleteWater",
        "description": "Delete a water record for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "_id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Water record deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/water/today": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get today's water stats",
        "operationId": "getTodayStats",
        "description": "Retrieve today's water consumption stats for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved today's water stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "waterRecords": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/water"
                      }
                    },
                    "totalWaterAmount": {
                      "type": "number",
                      "example": 2000
                    },
                    "dailyNorm": {
                      "type": "number",
                      "example": 2000
                    },
                    "percentage": {
                      "type": "number",
                      "example": 100
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/water/month/{date}": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get monthly water stats",
        "operationId": "getMonthlyStats",
        "description": "Retrieve mountly water consumption stats for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "YYYY-MM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved monthly water stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "date": {
                      "type": "string",
                      "format": "date",
                      "example": "2025-02-06T15:17:00.000Z"
                    },
                    "waterRecords": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/water"
                      }
                    },
                    "totalAmount": {
                      "type": "number",
                      "example": 7766
                    },
                    "dailyNorm": {
                      "type": "number",
                      "example": 2500
                    },
                    "percentage": {
                      "type": "number",
                      "example": 89
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/water/daily-norm": {
      "patch": {
        "tags": [
          "Water"
        ],
        "summary": "Update daily water norm",
        "operationId": "updateDailyNorm",
        "description": "Update daily norm for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "waterRate": {
                    "type": "number",
                    "example": 2500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Daily norm updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Daily norm updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "waterRate": {
                          "type": "number",
                          "example": 2500
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": 6012837263100000
                            },
                            "email": {
                              "type": "string",
                              "example": "john.doe@example.com"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not Found error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Not Found error"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Resource not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "InternalServerError"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An unexpected error occurred"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "water": {
        "type": "object",
        "required": [
          "userId",
          "waterVolume",
          "time"
        ],
        "properties": {
          "_id": {
            "description": "Water's id",
            "type": "string",
            "example": "65ca67e7ae7f10c88b598384"
          },
          "userId": {
            "description": "User's id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "waterVolume": {
            "description": "Volume of water consumed",
            "type": "number",
            "example": 2000
          },
          "date": {
            "description": "Date of water consumption",
            "type": "string",
            "format": "date-time",
            "example": "2021-08-09T00:00:00.000Z"
          },
          "time": {
            "description": "Time of water consumption",
            "type": "string",
            "example": "12:00"
          },
          "totalAmount": {
            "description": "Total amount of water consumed",
            "type": "number",
            "example": 2000
          },
          "dailyNorm": {
            "description": "Daily norm of water consumption",
            "type": "number",
            "example": 2000
          }
        }
      }
    }
  }
}